Add borders to screenshots
This patch adds 3px border to all the screenshots in the readme.
Markdown does not support in general image borders and it is a bit hacky
to add one that is compliant with all the markdown platforms so it is
easier to just include the border in the image content.